requirements.txt
django>=1.10
django-allauth
psycopg2
python-decouple
Pillow
markdown
